Moved Tileset out of Resource.
