+++ src/haven/SpriteLink.java
@@ -29,7 +29,6 @@ package haven;
 import java.util.*;
 import haven.Sprite.Factory;
 import haven.Sprite.Owner;
-import haven.Resource.Tileset;
 
 @Resource.LayerName("slink")
 public class SpriteLink extends Resource.Layer {
